Move drawing of repairs table to a utility function
