+++ panel-qc-viewer/public/javascripts/panel_qc.js
@@ -1,5 +1,6 @@
 
 import { plot_panel_qc } from './panel_qc_plot.js'
+import { draw_repairs_table } from './repairs_table.js'
 
 const form = document.querySelector("form");
 const log = document.querySelector("#log");
@@ -35,44 +36,13 @@ showPanelButton.addEventListener('click', async function () {
     document.getElementById("panel_info").innerHTML = output;
 
     // Now do the table
-    var cols = ["date_uploaded", "comment", "column_changed", "old_value", "new_value"];
+
     if (!isNaN(panel_number)) {
 	const repairs_table_response = await fetch('getPanelRepairs/'+panel_number);
-	const repairs_table = await repairs_table_response.json();
+	const repairs_table_info = await repairs_table_response.json();
 
 	var over_table = document.getElementById("panel_repairs_table");
-
-	while (over_table.firstChild) {
-	    over_table.removeChild(over_table.lastChild);
-	}
-
-	var table = document.createElement('TABLE');
-	table.border = '1';
-
-	var tableBody = document.createElement('TBODY');
-	table.appendChild(tableBody);
-
-	for (var i = 0; i < repairs_table.length+1; i++) {
-	    var tr = document.createElement('TR');
-	    tableBody.appendChild(tr);
-	    tr.border = '1'
-
-	    for (var j = 0; j < cols.length; j++) {
-		var td = document.createElement('TD');
-//		td.width = '75';
-		td.style.border = "1px solid #000"
-		if (i == 0) {
-		    td.appendChild(document.createTextNode([cols[j]]));
-		    td.style.textAlign = "center";
-		    td.style.borderBottomWidth = "2px"
-		}
-		else {
-		    td.appendChild(document.createTextNode(repairs_table[i-1][cols[j]]));
-		}
-		tr.appendChild(td);
-	    }
-	}
-	over_table.appendChild(table);
+	draw_repairs_table(repairs_table_info, over_table);
     }
     else {
 	output = "Input must be a number";
